Lessons/w13/demo-back/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==3.2.14
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
filelock==3.10.2
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.24.1
openpyxl==3.1.2
Pillow==9.4.0
platformdirs==3.1.1
pycodestyle==2.10.0
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2023.3
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
virtualenv==20.21.0
virtualenvwrapper-win==1.2.7
Werkzeug==2.2.2
